flask-api/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
python-dotenv
psycopg[binary]
gunicorn